@@ -4,5 +4,5 @@ namespace Domain.Models;
 
 public class User : IdentityUser
 {
-    public IEnumerable<Order> Orders { get; } = new List<Order>();
+    public IEnumerable<Order> Orders { get; } = [];
 }
